BeyondPPF-baseline/pointnet2/requirements.txt
open3d
omegaconf
hydra-core
tqdm